% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_info.R
\name{argument_validation}
\alias{argument_validation}
\title{Function that checks the format of arguments entered}
\usage{
argument_validation(stock_name, start_date, end_date)
}
\arguments{
\item{stock_name}{a sting that represent the symbol of the Company listed in NASDAQ}

\item{start_date}{a sting that represent the date}

\item{end_date}{a sting that represent the date}
}
\value{
TRUE if all the arguments are perfect else terminates the process
}
\description{
Function that checks the format of arguments entered
}
\details{
This function checks validity of stock name and if the date value entered is in the YYYY-MM-DD format
}
\author{
Rijin Baby
}
