% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_info.R
\name{check_file}
\alias{check_file}
\title{Function that search for data file}
\usage{
check_file(stock_name, start_date, end_date)
}
\arguments{
\item{stock_name}{a sting that represent the symbol of the Company listed in NASDAQ}

\item{start_date}{a sting that represent the date}

\item{end_date}{a sting that represent the date}
}
\value{
file path of the csv file
}
\description{
Function that search for data file
}
\details{
This function initially checks if the file with stock details is present in the directory or not.
If the file is found in the directory the file path is returned else this function download the file
from the NASDAQ website and return the file path. In case of no data present in the file the function terminates the process
}
\seealso{
\code{download.file}
}
\author{
Rijin Baby
}
