% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_info.R
\name{get_data}
\alias{get_data}
\title{Function that search and download the data file}
\usage{
get_data(stock_name)
}
\arguments{
\item{stock_name}{a string that represent the symbol of the Company listed in NASDAQ}
}
\value{
temporary file path of the csv file
}
\description{
Function that search and download the data file
}
\details{
This function download the stock data fro the NASDAQ website to a temporary directory in the system
and return the file path for better memory management.
}
\seealso{
\code{download.file}
}
\author{
Rijin Baby
}
