% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_info.R
\name{stock_details}
\alias{stock_details}
\title{Function that displays and plot the stock data. By default the function gives 10 years data for the stock if the start and end dates are not specified}
\usage{
stock_details(
  stock_name,
  start_date = Sys.Date() - years(10),
  end_date = Sys.Date()
)
}
\arguments{
\item{stock_name}{a sting that represent the symbol of the Company listed in NASDAQ}

\item{start_date}{a sting that represent the date; By default 10 years from current date is considered}

\item{end_date}{a sting that represent the date; Current date is taken as default if no value is provided}
}
\value{
Data and plot
}
\description{
Function that displays and plot the stock data. By default the function gives 10 years data for the stock if the start and end dates are not specified
}
\details{
The function initially cleans the entered argumets and check their validity. Then update the data file and finally
dispalys the data for the analysis
}
\seealso{
\code{plot_ly}\code{layout}


}
\author{
Rijin Baby
}
