% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_info.R
\name{stock_details}
\alias{stock_details}
\title{Function that displays and plot the stock data}
\usage{
stock_details(stock_name, start_date, end_date = Sys.Date())
}
\arguments{
\item{stock_name}{a sting that represent the symbol of the Company listed in NASDAQ}

\item{start_date}{a sting that represent the date}

\item{end_date}{a sting that represent the date; Current date is taken as default if no value is provided}
}
\value{
Data and plot
}
\description{
Function that displays and plot the stock data
}
\details{
The function initially cleans the entered argumets and check their validity. Then collect the data file,
dispalys the data for the analysis
}
\seealso{
\code{ggplot} \code{aes} \code{labs} \code{geom_line}


}
\author{
Rijin Baby
}
